% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drift.simulator.R
\name{drift.sim}
\alias{drift.sim}
\title{Simulation drift with a Wolbachia population}
\usage{
drift.sim(p, N, s, g, fit, r)
}
\arguments{
\item{p}{Initial frequency of the population to be modeled}

\item{N}{Population size per generation}

\item{s}{Number of individuals to be sampled at each generation}

\item{fit}{Fitness of the population to be modeled. Zero (0) gives no fitness advantage.}

\item{r}{Number of replicates}

\item{Total}{generation to be modeled}
}
\value{
A figure with different results.
}
\description{
Drift simulation taking in account a population with to different phenotypes
}
\author{
Elves H Duarte
}

