% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysing_qpcr_data.R
\name{fold.calc}
\alias{fold.calc}
\title{Analyze qPCR data in R}
\usage{
fold.calc(samples, genes, Cts, standev = standev, sd.rm = TRUE, ref, data)
}
\arguments{
\item{samples}{A column containing your samples name}

\item{genes}{A column with your reference and control genes}

\item{Cts}{Your Ct values}

\item{standev}{The maximum standard deviation among your technical replicates}

\item{sd.rm}{A logical parameter to decide whether the  outlayer sample will be removed}

\item{ref}{The reference samples}

\item{data}{the name of your data frame}
}
\value{
Return a five-column data frame
}
\description{
It takes the your date and by computing accordingly to Pfaffl method it return the fold changes
}
\author{
Elves H Duarte
}

