% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phred.R
\name{phred}
\alias{phred}
\title{Understand Phred quality}
\usage{
phred(val, get = "error")
}
\arguments{
\item{val}{A error (range 0 to 1) or quality (greather than 0) value to be converted whether to Phred quality or error probability.}

\item{get}{Specify to what you want to convert the value: 'error' convert a Phred value to error and 'quality' convert a desired error probability to Phred quality}
}
\value{
A error probability of a Phred quality or the Phred quality associated with a error.
}
\description{
Convert between Phred quality index and error probablity
}
\author{
Elves H Duarte
}

