% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos2gene.R
\name{pos2gene}
\alias{pos2gene}
\title{Annotate Wolbachia positions}
\usage{
pos2gene(x)
}
\arguments{
\item{x}{A list of positions to be annotated.}
}
\value{
Retur a list of genes (or not) of Wolbachia locus.
}
\description{
By using a bed file each position of Wolbahcia is annotated
}
\author{
Elves H Duarte
}

