% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatter plots and median or mean bars}
\usage{
scatter(x, data, fun = "median", main = "", sub = "", xlim = NULL,
  ylim = NULL, axes = TRUE, ann = par("ann"), col = par("col"), ...)
}
\arguments{
\item{x}{The values to be plotted}

\item{fun}{Choose which statistic (medium or mean) will be added}

\item{y}{Categorical variable containing the different groups}
}
\value{
A plot based in 1-D scatter plot of the base package
}
\description{
This function creats a scatter plot and add groups bar representing median or mean values
}
\author{
Elves H Duarte
}

